DatabaseAnalysis via LoadOnDemandMappedTable now
Rethought out the concept of caching with analyses. Now analysis fields
will be stored via LoadOnDemandMappedTable and its subclass
DatabaseAnalysisResultsTable. Each analysis field value will be cached
and loaded individually. LoadOnDemandMappedTable also encapsulates
the more general concept of an entry that is related via a one-to-one
relationship with another class's entries, but whose field values are
either too large to load all at once, or should be computed on the fly
and never stored. This table type facilitates this.